+++ src/bin/statinamic-setup.js
@@ -3,20 +3,24 @@ import npmInstallPackage from "npm-install-package"
 import { peerDependencies as peerDeps } from "../../package.json"
 
 console.log("Installing required dependencies...")
+const depsMap = Object.keys(peerDeps)
+  .reduce(
+    (deps, dep) => {
+      deps.push(`${ dep }@${ peerDeps[dep] }`)
+      return deps
+    },
+    []
+  )
 npmInstallPackage(
-  Object.keys(peerDeps).reduce((deps, dep) => {
-    deps.push(`${ dep }@${ peerDeps[dep] }`)
-    return deps
-  }, []),
-  { saveDev: true },
-  err => {
-    if (err) throw err
+  depsMap,
+  {
+    saveDev: true,
+  },
+  (err) => {
+    if (err) {
+      throw err
+    }
+
+    console.log("Required dependencies successfully installed.")
   }
 )
-console.log("Required dependencies successfully installed.")
-
-// TODO copy boilerplate
-// TODO add npm scripts
-
-// TODO ask/add some metadata (twitter, github, analytics) ?
-// TODO add babel config ?